+++ src/app/orders/services/order.service.ts
@@ -12,7 +12,6 @@ export class OrderService {
   private orders: Array<Order> = JSON.parse(this.localStorageService.getItem('orders')) || [];
 
   getOrders(): Array<Order> {
-    console.log(this.orders);
     return this.orders;
   }