+++ src/app/shared/pipes/order-by.pipe.ts
@@ -8,8 +8,6 @@ export class OrderByPipe implements PipeTransform {
   transform(array: any[], field: string, dir: string): any[] {
     let val: number;
     if ( dir === 'true' ) { val = -1; } else { val = 1; }
-    console.log(val);
-    console.log(dir);
     if (!Array.isArray(array)) {
       return array;
     }